+++ src/helpers/stringUtil.js
@@ -2,6 +2,7 @@ import { defineMessages } from 'react-intl';
 import { TransactionStatus } from 'constants';
 
 import { getIntlProvider } from './i18nUtil';
+import { shortenText } from './utility';
 
 const strings = defineMessages({
   strPendingMsg: {
@@ -15,6 +16,10 @@ const strings = defineMessages({
   strFailMsg: {
     id: 'str.fail',
     defaultMessage: 'Fail',
+  },
+  strYou: {
+    id: 'str.you',
+    defaultMessage: 'You',
   },
 });
 export function getStatusString(txStatus, intl) {
@@ -30,4 +35,10 @@ export function getStatusString(txStatus, intl) {
       return formatMessage(strings.strFailMsg);
     }
   }
+}
+
+export function getAddressName(account, intl, address, name) {
+  return (account && account.toLowerCase() === address && intl.formatMessage(strings.strYou))
+    || name
+    || shortenText(address, 6);
 }